Add macro validation engine test
