+++ validation/src/test/scala/hmda/validation/rules/lar/macro/MacroTestData.scala
@@ -1,0 +1,16 @@
+package hmda.validation.rules.lar.`macro`
+
+import hmda.model.fi.lar.LoanApplicationRegister
+import hmda.model.util.FITestData._
+import hmda.parser.fi.lar.LarCsvParser
+
+object MacroTestData {
+
+  val lars: Array[LoanApplicationRegister] = {
+    fiCSV.split("\n")
+      .tail.map(line => LarCsvParser(line))
+      .filter(x => x.isRight)
+      .map(x => x.right.get)
+  }
+
+}
